requirements.txt
opencv-python==4.8.1.78
pyserial==3.5
numpy==1.24.3
